Replace id3tag with taglib-ruby
I want to be able to extract the track duration and I'm fairly confident I
can't do that with id3tag, but I can with taglib-ruby.

The taglib-ruby gem also appears to handle the case where a single MP3 has two
"TDRC" ID3 frames with the same value ("2011" in my case). I'd previously had
to work around this because id3tag raised an exception.